Create READ.ME
At the company's development department, we have been struggling a lot lately with our production applications and that has caused constant downtimes and complaints from our clients. We have identified the problem and we figured out that we are not handling our logs properly, which in turn causes the disk to get full and it eventually crashes the entire server with the applications in it. We've partnered with the SRE team and they have helped us create a plan that will allow us to solve the problems, have better control over the logs and gain visibility through notifications. We would like you to implement it using shell scripts since we lack expertise with Linux and scripting in general.

We will provide you with the templates/files that the SRE team handed us. At a very high level, we want to implement multiple small scripts rather than a gigantic one; this will ease maintainability as well as readability for the team. Each script will have a unique responsibility. One script will generate mock logs to test the scenarios, the next one will compress them, the third one will generate a report and the final one will send notifications based on the report. As I said in the beginning, the initial structure and plan from the SRE will be handed to you in the form of "empty scripts with comments". You will have to go through each script, write the behavior described in the script, and test it properly. The implementation is completely up to you as long as it complies with the templates provided by the SRE team. We're counting on you!